requirements.txt
torch
torchvision
sklearn
pandas
pickle
opencv-python
tqdm
torch-geometric==2.5.3
scikit-image==0.22.0
numpy==1.26.4
scipy==1.11.4
networkx==3.2.1
pillow==9.5.0
imageio==2.33.1
tifffile==2023.12.9
packaging==21.3
lazy_loader==0.3
pyparsing==3.1.1